3619: Rename AttributeNameWithDoubleQuotationMarksIssueGenerator files
